requirements.txt
# Django/DRF core
Django==5.2.2
djangorestframework==3.16.0
djangorestframework-simplejwt==5.4.0
drf-spectacular==0.28.0

# Database & config
psycopg2-binary==2.9.10
dj-database-url==2.3.0

# Security & middleware
django-cors-headers==4.6.0
django-ratelimit==4.1.0

# Async & clients
celery==5.4.0
redis==5.1.1
httpx==0.28.1

# Chain
web3==7.2.0
eth-account>=0.13.1
eth-utils==2.3.1

# Utils
python-dotenv==1.0.1
PyJWT==2.10.1
gunicorn==22.0.0
python-json-logger==2.0.7
django-redis==5.4.0
# Email handled via Omnisend API (omnisend_service.py)

# Testing
pytest==8.3.2
pytest-django==4.9.0
respx==0.21.1
freezegun==1.5.1
coverage==7.6.4